# dashboard/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Memory Observer Dashboard API
  description: |
    REST API for the Memory Observer Dashboard, providing access to workspaces, plans, agents, and events
    from the Project Memory MCP Server.
    
    ## Overview
    
    The Memory Observer Dashboard API allows you to:
    - View and manage MCP workspaces
    - Browse, create, and manage project plans
    - View and deploy agent templates
    - Subscribe to real-time MCP events
    - Search across all workspaces and plans
    
    ## Authentication
    
    Currently, the API runs locally and does not require authentication.
    Future versions may support authentication for remote access.
    
  version: 0.1.0
  contact:
    name: Project Memory MCP
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Local Development Server
...
paths:
...
  /api/plans/{workspaceId}/{planId}:
    get:
      tags:
        - Plans
      summary: Get plan details
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/PlanId'
      responses:
        '200':
          description: Plan details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags:
        - Plans
      summary: Delete a plan
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/PlanId'
        - name: permanent
          in: query
          description: Permanently delete the plan files
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Plan deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  planId:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/plans/{workspaceId}/{planId}/archive:
    post:
      tags:
        - Plans
      summary: Archive a plan
      description: Move a plan to archived status
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/PlanId'
      responses:
        '200':
          description: Plan archived
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  planId:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/plans/{workspaceId}/{planId}/duplicate:
    post:
      tags:
        - Plans
      summary: Duplicate a plan
      description: Create a copy of a plan as a template
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/PlanId'
      responses:
        '201':
          description: Plan duplicated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  newPlanId:
                    type: string
                  newPlan:
                    $ref: '#/components/schemas/Plan'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/plans/{workspaceId}/import:
    post:
      tags:
        - Plans
      summary: Import a plan from file
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - filePath
              properties:
                filePath:
                  type: string
                  description: Path to the markdown file to import
      responses:
        '201':
          description: Plan imported
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  planId:
                    type: string
                  plan:
                    $ref: '#/components/schemas/Plan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      description: Workspace ID
      schema:
        type: string
        example: ds_my-project-abc123
    PlanId:
      name: planId
      in: path
      required: true
      description: Plan ID
      schema:
        type: string
        example: plan_abc123_xyz789
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Resource not found
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid request parameters
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: An unexpected error occurred
  schemas:
...
    PlanSummary:
      type: object
      properties:
        id:
          type: string
          example: plan_abc123_xyz789
        workspaceId:
          type: string
        title:
          type: string
          example: Add Authentication
        priority:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [active, paused, completed, archived, failed]
        category:
          type: string
          enum: [feature, bug, change, analysis, debug, refactor, documentation]
        currentAgent:
          type: string
          nullable: true
        stepCount:
          type: integer
        completedStepCount:
          type: integer
        updatedAt:
          type: string
          format: date-time
    Plan:
      allOf:
        - $ref: '#/components/schemas/PlanSummary'
        - type: object
          properties:
            description:
              type: string
            currentPhase:
              type: string
            createdAt:
              type: string
              format: date-time
            steps:
              type: array
              items:
                $ref: '#/components/schemas/PlanStep'
            agentSessions:
              type: array
              items:
                $ref: '#/components/schemas/AgentSession'
            lineage:
              type: array
              items:
                $ref: '#/components/schemas/LineageEntry'
    PlanStep:
      type: object
      properties:
        phase:
          type: string
        task:
          type: string
        status:
          type: string
          enum: [pending, active, done, blocked]
        notes:
          type: string
          nullable: true
    AgentSession:
      type: object
      properties:
        agentType:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [active, completed, failed]
        summary:
          type: string
          nullable: true
    LineageEntry:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        fromAgent:
          type: string
        toAgent:
          type: string
        reason:
          type: string
...